requirements.txt
gradio==4.40.0
pydantic==2.8.2
openai>=1.30.0
PyYAML==6.0.2
